data_analyzer/requirement.txt
Django==2.2.7
djangorestframework == 3.14.0
pandas==2.0.0
